refactor
refactor component naming laws and reafactor homepage
